llm-router/pyproject.toml
[tool.poetry]
name = "llm-router"
version = "1.0.0"
description = "Lightweight LLM provider routing service"
authors = ["Maestro Platform <dev@maestro.ai>"]
readme = "README.md"
packages = [{include = "llm_router", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0"
redis = "^5.0.0"
openai = "^1.10.0"
anthropic = "^0.39.0"
google-generativeai = "^0.3.2"
httpx = "^0.26.0"
claude-agent-sdk = "^0.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.12.0"
ruff = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
